Some refactoring
Rename some classes, remove useless comments, catch key error and some
others improvements.

+++ src/me/shadorc/spinvader/entity/Entity.java
@@ -17,8 +17,8 @@ public interface Entity {
 		MONEY, LIFE;
 	}
 
-	public float getX();
-	public float getY();
+	public int getX();
+	public int getY();
 	public float getLife();
 
 	public Rectangle getHitbox();
